Issue #2044 Create separate ShutdownServiceImpl, extract all functionality related to DTS preferences management to PreferenceService
