+++ data-transfer-service/src/main/java/com/epam/pipeline/dts/sync/SyncConfiguration.java
@@ -16,53 +16,14 @@
 
 package com.epam.pipeline.dts.sync;
 
-import com.epam.pipeline.dts.common.service.CloudPipelineAPIClient;
 import com.epam.pipeline.dts.configuration.CommonConfiguration;
-import com.epam.pipeline.dts.sync.model.AutonomousDtsDetails;
-import org.apache.commons.lang3.StringUtils;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.SpringBootConfiguration;
-import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Import;
 
-import java.net.InetAddress;
-import java.net.UnknownHostException;
-import java.util.Optional;
-
 @SpringBootConfiguration
-@ComponentScan(basePackages = "com.epam.pipeline.dts.sync.impl")
+@ComponentScan(basePackages = "com.epam.pipeline.dts.sync")
 @Import({CommonConfiguration.class})
 public class SyncConfiguration {
 
-    @Bean
-    public AutonomousDtsDetails autonomousDtsDetails(final @Value("${dts.name}") String dtsName,
-                                                     final CloudPipelineAPIClient apiClient) {
-        final String resolvedDtsName = tryBuildDtsName(dtsName);
-        return new AutonomousDtsDetails(resolvedDtsName, apiClient);
-    }
-
-
-    private String tryBuildDtsName(final String preconfiguredDtsName) {
-        final String dtsName = Optional.ofNullable(preconfiguredDtsName)
-            .filter(StringUtils::isNotBlank)
-            .orElseGet(this::tryExtractHostnameFromEnvironment);
-        if (StringUtils.isBlank(dtsName)) {
-            throw new IllegalStateException("Unable to build DTS name!");
-        }
-        return dtsName;
-    }
-
-    private String tryExtractHostnameFromEnvironment() {
-        try {
-            return Optional.ofNullable(InetAddress.getLocalHost())
-                .map(InetAddress::getCanonicalHostName)
-                .filter(StringUtils::isNotEmpty)
-                .map(StringUtils::strip)
-                .map(StringUtils::lowerCase)
-                .orElse(StringUtils.EMPTY);
-        } catch (UnknownHostException e) {
-            return StringUtils.EMPTY;
-        }
-    }
 }
