--- data-transfer-service/src/main/java/com/epam/pipeline/dts/sync/model/AutonomousDtsDetails.java
@@ -1,27 +1,0 @@
-/*
- * Copyright 2017-2021 EPAM Systems, Inc. (https://www.epam.com/)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.epam.pipeline.dts.sync.model;
-
-import com.epam.pipeline.dts.common.service.CloudPipelineAPIClient;
-import lombok.Value;
-
-@Value
-public class AutonomousDtsDetails {
-
-    private String dtsName;
-    private CloudPipelineAPIClient apiClient;
-}
