+++ data-transfer-service/src/main/java/com/epam/pipeline/dts/sync/service/impl/ShutdownServiceImpl.java
@@ -20,24 +20,21 @@ import com.epam.pipeline.dts.sync.service.PreferenceService;
 import com.epam.pipeline.dts.sync.service.ShutdownService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.scheduling.annotation.EnableScheduling;
-import org.springframework.scheduling.annotation.Scheduled;
+import org.springframework.context.ConfigurableApplicationContext;
 import org.springframework.stereotype.Service;
-
 
 @Service
 @RequiredArgsConstructor
 @Slf4j
-@EnableScheduling
-public class DtsSynchronizationService {
+public class ShutdownServiceImpl implements ShutdownService {
 
+    private final ConfigurableApplicationContext context;
     private final PreferenceService preferenceService;
-    private final ShutdownService shutdownService;
 
-    @Scheduled(fixedDelayString = "${dts.sync.poll:60000}")
-    public void synchronizePreferences() {
-        if (preferenceService.isShutdownRequired()) {
-            shutdownService.shutdown();
-        }
+    @Override
+    public void shutdown() {
+        log.info("Shutdown will be preformed.");
+        preferenceService.clearShutdownFlag();
+        context.close();
     }
 }
